+++ packages/thumbprint-react/components/Radio/index.tsx
@@ -126,6 +126,11 @@ export interface RadioProps {
      * Determines how the radio button input will be vertically aligned relative to `props.children`.
      */
     radioVerticalAlign?: 'top' | 'center';
+    /**
+     * Determines the value that will be submitted if the radio is selected. The default value is
+     * `'on'`.
+     */
+    value?: string | string[] | number;
 }
 
 export default function Radio({
@@ -141,6 +146,7 @@ export default function Radio({
     onChange,
     onKeyDown = (): void => {},
     radioVerticalAlign = 'center',
+    value,
 }: RadioProps): JSX.Element {
     const uiState = getUIState({ isChecked, isDisabled, hasError });
 
@@ -173,6 +179,7 @@ export default function Radio({
                 disabled={isDisabled}
                 data-test={dataTest}
                 required={isRequired}
+                {...(value ? { value } : {})}
             />
 
             <svg